+++ config/settings.yaml
@@ -25,6 +25,9 @@ services:
   Pimcore\Bundle\StudioBackendBundle\Setting\Service\BundlesServiceInterface:
     class: Pimcore\Bundle\StudioBackendBundle\Setting\Service\BundlesService
 
+  Pimcore\Bundle\StudioBackendBundle\Setting\Service\ElementServiceInterface:
+    class: Pimcore\Bundle\StudioBackendBundle\Setting\Service\ElementService
+
   #
   # Hydrators
   #
