+++ src/Document/Service/DocumentService.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Pimcore\Bundle\StudioBackendBundle\Document\Service;
 
-use Pimcore\Bundle\StaticResolverBundle\Lib\ConfigResolverInterface;
 use Pimcore\Bundle\StaticResolverBundle\Models\Document\DocumentServiceResolverInterface;
 use Pimcore\Bundle\StaticResolverBundle\Models\Element\ServiceResolverInterface;
 use Pimcore\Bundle\StudioBackendBundle\DataIndex\Query\DocumentQueryInterface;
@@ -21,11 +20,9 @@ use Pimcore\Bundle\StudioBackendBundle\DataIndex\Request\ElementParameters;
 use Pimcore\Bundle\StudioBackendBundle\DataIndex\SearchIndexFilterInterface;
 use Pimcore\Bundle\StudioBackendBundle\DataIndex\Service\DocumentSearchServiceInterface;
 use Pimcore\Bundle\StudioBackendBundle\Document\Event\PreResponse\DocumentEvent;
-use Pimcore\Bundle\StudioBackendBundle\Document\Event\PreResponse\DocumentTypeEvent;
 use Pimcore\Bundle\StudioBackendBundle\Document\Schema\Document;
 use Pimcore\Bundle\StudioBackendBundle\Document\Schema\DocumentAddParameters;
 use Pimcore\Bundle\StudioBackendBundle\Document\Schema\DocumentDetail;
-use Pimcore\Bundle\StudioBackendBundle\Document\Schema\DocumentType;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ElementExistsException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ForbiddenException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\InvalidElementTypeException;
@@ -40,7 +37,6 @@ use Pimcore\Bundle\StudioBackendBundle\Util\Trait\UserPermissionTrait;
 use Pimcore\Model\Document as DocumentModel;
 use Pimcore\Model\UserInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use function is_array;
 use function sprintf;
 
 /**
@@ -53,7 +49,6 @@ final readonly class DocumentService implements DocumentServiceInterface
 
     public function __construct(
         private CreateServiceInterface $createService,
-        private ConfigResolverInterface $configResolver,
         private DataServiceInterface $dataService,
         private DocumentSearchServiceInterface $documentSearchService,
         private DocumentServiceResolverInterface $documentServiceResolver,
@@ -163,21 +158,6 @@ final readonly class DocumentService implements DocumentServiceInterface
         }
 
         return $document;
-    }
-
-    public function getDocumentTypes(): array
-    {
-        $types = [];
-        $config = $this->configResolver->getSystemConfiguration('documents');
-        if (isset($config['type_definitions']['map']) && is_array($config['type_definitions']['map'])) {
-            foreach ($config['type_definitions']['map'] as $key => $definition) {
-                $type = new DocumentType($key);
-                $this->eventDispatcher->dispatch(new DocumentTypeEvent($type), DocumentTypeEvent::EVENT_NAME);
-                $types[] = $type;
-            }
-        }
-
-        return $types;
     }
 
     /**
