+++ src/Document/Service/DocumentServiceInterface.php
@@ -16,7 +16,6 @@ namespace Pimcore\Bundle\StudioBackendBundle\Document\Service;
 use Pimcore\Bundle\StudioBackendBundle\DataIndex\Request\ElementParameters;
 use Pimcore\Bundle\StudioBackendBundle\Document\Schema\DocumentAddParameters;
 use Pimcore\Bundle\StudioBackendBundle\Document\Schema\DocumentDetail;
-use Pimcore\Bundle\StudioBackendBundle\Document\Schema\DocumentType;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ElementExistsException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ElementSavingFailedException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ForbiddenException;
@@ -66,9 +65,4 @@ interface DocumentServiceInterface
      * @throws ForbiddenException|NotFoundException
      */
     public function getDocumentElementByPath(UserInterface $user, string $path): DocumentModel;
-
-    /**
-     * @return DocumentType[]
-     */
-    public function getDocumentTypes(): array;
 }
